Prohibit restart of docker or containerd services
When doing package upgrades, services get restarted by default. For most
services this is not an issue. However, restarting Docker or containerd
will leave the system in a state where dpkg doesn't work, since a Docker
or containerd restart will terminate the container running the apt-get
upgrade in the middle of the process.

This commit adds a [policy-rc.d] script which prohibits stops or
restarts of the docker and containerd system services. The package
upgrade script will install the provided policy-rc.d script as
/usr/sbin/policy-rc.d before doing package upgrades.

The policy-rc.d script ensures that /var/run/reboot-required exists if
either Docker or containerd was upgraded (as indicated by the call to
the policy-rc.d script for the service stop or restart). By creating
/var/run/reboot-required the code that triggers a node reboot after the
package updates is triggered when Docker and/or containerd are upgraded.

Resolves #9

[policy-rc.d]: https://people.debian.org/~hmh/invokerc.d-policyrc.d-specification.txt