Implement optional apt-get update during maintenance window
This provides a short-term fix for #14 for users who aren't particularly
interested in having deterministic upgrades throughout the week while we
determine how to address the root cause.

This commit introduces an optional command line flag `-u` for
`/scripts/run.sh` which triggers a `apt-get update` at the time the
script is executed to upgrade packages on a host.