Catch exceptions thrown by `pushadd_to_gateway()`
We catch exceptions thrown by `pushadd_to_gateway()` in
`update_exporter_deb.py` and present the errors in a nicer format.

Note that we still exit the Python script with an error, but since we
ignore errors in `update-packages.sh` this should still ensure that
failures when communicating with the push gateway don't fail the
maintenance job.